// src/app/Interfaces/ipelement.ts
export interface IPElement {
    atomicNumber: number;
    symbol: string;
    name: string;
    atomicMass: any;
    standardState: string;
    yearDiscovered: number;
    cpkHexColor: string;
    period: number;
    group: number;
    meltingPoint: number;
    boilingPoint: number;
    atomicRadius: number;
    electronegativity: string;
    electronicConfiguration: string;
    density: number;
    groupBlock: string;
    color?: string;
}